+++ src/conpty/mod.rs
@@ -1,4 +1,4 @@
-use std::{fs::File, sync::Arc};
+use std::{fs::File, process::Command, sync::Arc};
 
 mod file;
 pub mod gather;
@@ -22,15 +22,23 @@ pub const PTY_INSPECT_READ: u8 = 0x0;
 pub const PTY_INSPECT_WRITE: u8 = 0x1;
 pub const SLOT_PTY_BIN: &str = "event_slot_pty_file";
 pub const PTY_FLAG_ENABLE_BLOCK: u32 = 0x00000001;
+#[cfg(not(test))]
+pub const PTY_EXEC_DATA_SIZE: usize = 2048;
+#[cfg(test)]
+pub const PTY_EXEC_DATA_SIZE: usize = 5;
 
 const WS_MSG_TYPE_PTY_ERROR: &str = "PtyError";
 const WS_MSG_TYPE_PTY_EXEC_CMD: &str = "PtyExecCmd";
+const WS_MSG_TYPE_PTY_EXEC_CMD_STREAM: &str = "PtyExecCmdStream";
 const WS_MSG_TYPE_PTY_START: &str = "PtyStart";
 const WS_MSG_TYPE_PTY_STOP: &str = "PtyStop";
 const WS_MSG_TYPE_PTY_RESIZE: &str = "PtyResize";
 const WS_MSG_TYPE_PTY_INPUT: &str = "PtyInput";
 const WS_MSG_TYPE_PTY_READY: &str = "PtyReady";
 const WS_MSG_TYPE_PTY_OUTPUT: &str = "PtyOutput";
+
+type PtyExecCallback = Box<dyn Fn(u32, bool, Option<i32>, Vec<u8>)>;
+type PtyResult<T> = Result<T, String>;
 
 pub trait PtyAdapter {
     fn openpty(
@@ -39,13 +47,19 @@ pub trait PtyAdapter {
         cols: u16,
         rows: u16,
         flag: u32,
-    ) -> Result<Arc<dyn PtyBase + Send + Sync>, String>;
+    ) -> PtyResult<Arc<dyn PtyBase + Send + Sync>>;
 }
 pub trait PtyBase {
-    fn resize(&self, cols: u16, rows: u16) -> Result<(), String>;
-    fn get_reader(&self) -> Result<File, String>;
-    fn get_writer(&self) -> Result<File, String>;
-    fn get_pid(&self) -> Result<u32, String>;
-    fn inspect_access(&self, path: &str, access: u8) -> Result<(), String>;
-    fn execute(&self, f: &dyn Fn() -> Result<Vec<u8>, String>) -> Result<Vec<u8>, String>;
+    fn resize(&self, cols: u16, rows: u16) -> PtyResult<()>;
+    fn get_reader(&self) -> PtyResult<File>;
+    fn get_writer(&self) -> PtyResult<File>;
+    fn get_pid(&self) -> PtyResult<u32>;
+    fn inspect_access(&self, path: &str, access: u8) -> PtyResult<()>;
+    fn execute(&self, f: &dyn Fn() -> PtyResult<Vec<u8>>) -> PtyResult<Vec<u8>>;
+    fn execute_stream(
+        &self,
+        cmd: Command,
+        cb: Option<PtyExecCallback>,
+        timeout: Option<u64>,
+    ) -> PtyResult<()>;
 }